Use floor() insstead of floorf()

+++ src/Cockpit/panel.cxx
@@ -1118,7 +1118,7 @@ FGTextLayer::Chunk::getValue () const
       break;
     case DOUBLE_VALUE:
       double d = _offs + _node->getFloatValue() * _mult;
-      if (_trunc)  d = (d < 0) ? -floorf(-d) : floorf(d);
+      if (_trunc)  d = (d < 0) ? -floor(-d) : floor(d);
       sprintf(_buf, _fmt.c_str(), d);
       break;
     }
